Add BuildEvents to Ant
The basic idea is to keep the core build engine "clean" and free of any
presentation logic, and to make it easier to extend Ant with other features
without cluttering up the core. To do this, I've defined a BuildListener
interface and added an "addBuildListener" method to Project that can be
used to register listener objects. Listeners could be implemented to
generate reports, send out emails when the build is complete, create a
bill of materials, etc...

The only new functionality visible to the end-user is a "-listener" option
on the command line that will let you specify the name of a class. An
instance of this class will be added as a listener to the project. I've
included a listener that will generate an XML log file, which you can use
by typing the command below.

build -listener org.apache.tools.ant.XmlLogger

PR:
Obtained from:
Submitted by:	Matt Foemmel <mpfoemme@ThoughtWorks.com>
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/ant@267694 13f79535-47bb-0310-9956-ffa450edef68