Move to task level logging
This patch removes the concept of currentTarget and currentTask from the
BuildEvents code. To facilitate this I have moved logging to the task
level which now passes a task pointer to the project's log method.

Task level logging may also allow for more fine grained control of logging
in the future.

I have left the Project's log methods public to allow people's custom
tasks to continue to work. In the future these can become private.


git-svn-id: http://svn.apache.org/repos/asf/ant@267740 13f79535-47bb-0310-9956-ffa450edef68