New build procedure for Unix based systems
This is the update for the Unix based build scripts. The changes are the
following

Automatic bootstrapping if lib/ant.jar is mising or the bin files
are not present

Removed inclusion of the $HOME/.antrc in the bootstrap.sh although
it is still used in the ant shell script.

Lined up Windows and Unix scripts, including echo statements so it is easier
to make changes affecting both envionments in the future.

Added chmod calls to the install targets so the ant shell scripts are
executable.


git-svn-id: http://svn.apache.org/repos/asf/ant@267777 13f79535-47bb-0310-9956-ffa450edef68