Restore antRun.bat for Win9x users.
This should allow the exec task to execute on Win9x. I have changed Glen's
patch to test for platfrom != NT rather than == 98 so it will
support 95.

NOTE: There is an important limitation. Only 9 arguments can be
passed to the command being exec'ed when the antRun.bat file is used. This
should only occur when the directory is not . and the platform is not NT.
This was the original reason this was changed (and antRun.bat) deleted.


Submitted by:	Glen Stampoultzis <trinexus@one.net.au>


git-svn-id: http://svn.apache.org/repos/asf/ant@267789 13f79535-47bb-0310-9956-ffa450edef68