Changes to EJBJar task
Make the descriptordir attribute optional.  If it is not there, then it
will default to the original functionality by looking srcdir.

Added an optional basejarname attribute.  If this exists, then there is
no need to prefix your descriptors with the jar name.  If it is not
used, then it will default to using the original naming convention.

Added a optional weblogicdtd attribute for the location of
the weblogic DTD file.  If this is not there, then the original
functionality will be  executed (uses the classpath resource
/weblogic/ejb/deployment/xml/ejb-jar.dtd as the location of the
dtd file).

Added a weblogictoplink element to the ejbjar task.  This
subtask allows you to build TOPLink for WebLogic enabled ejb jar
files. This subclasses the weblogic element so it uses the same
attributes with
the following additions:
    toplinkdescriptor  - this is the name of the xml file to be used by
    (required)           TOPLINK/ejbc.  This does not require the
                         naming prefix as the naming standard
                         (suggested by The ObjectPeople/WebGain) for
                         these files is incompatible with the current
                         naming convention of the ejbjar task.

    toplinkdtd         - the location of the TOPLink DTD file.  If this
    (optional)           is not provided, the web address is used
                         (Unfortunately, this cannot be picked up like
                         the weblogic DTD)

Submitted by:	John Hall <jhall@solant.com>


git-svn-id: http://svn.apache.org/repos/asf/ant@267977 13f79535-47bb-0310-9956-ffa450edef68