First part of the scanDir refactoring. This is just a starter to show what I want to accomplish.
All tasks using a scanDir method to find the source files that are
newer than there targets should get a similar treatment as <copy>.

The major benefit is not to have the functionality in one place but
the added flexibility you get, when you allow the user to specify a
custom FileNameMapper to a task (not there yet).

Nothing (especially not the names) is carved into stone of course.


git-svn-id: http://svn.apache.org/repos/asf/ant@268190 13f79535-47bb-0310-9956-ffa450edef68