Made scripts a little more consistent, they should now work with IBM's JDK on AIX out of the box.
Ant will now prepend CLASSPATH to the jars in ANT_HOME/lib like it
always did for Windows - if this is not the correct order, change it
here and in the batch file. We need to be consistent.

Make sure :: doesn't slip into the CLASSPATH when adding the library
jars.


git-svn-id: http://svn.apache.org/repos/asf/ant@269015 13f79535-47bb-0310-9956-ffa450edef68