This is a major change. :-)
It introduces the concept of a TaskContainer to allow a task to contain
other tasks. This allows Task composition
It introduces a <parallel> task for multithreading support. There is
also a <sequential> task.
It reworks System.out management to handle all task generated output
and route it through the Ant event system. This handles multithreaded
output.

This is a major rework to the patch originally submitted by Thomas. I
have taken a different route for the output management, in particular.

Based on patch by Thomas Christen <chr@active.ch>


git-svn-id: http://svn.apache.org/repos/asf/ant@269371 13f79535-47bb-0310-9956-ffa450edef68