Skip SQLExecTest if the testcase and the <sql> implementation would be loaded by different classloaders.
package private access is only allowed if the classes belong to the
same "runtime package", which is the fully qualified package name plus
the defining classloader.  Enough reading of JLS and VMSPEC for me for
today 8-)

We probably could code around the problem by placing a subclass of
SQLExec into the same classloader as SQLExecTest and invoke methods on
this subclass.


git-svn-id: http://svn.apache.org/repos/asf/ant@270618 13f79535-47bb-0310-9956-ffa450edef68