Added the TaskFactory and modified Project.
This is not organized as a 'project' - the files here needs to replace
the ones in the main tree. I'll probably import the rest of the core
and make it a standalone replacement, so it's easy to review.

The goals are to:
- make ant more 'embeddable'
- integrate it better in other tools
- support SAX2 and namespaces
- support pluggable behavior for the XML reader
- support pluggable task factories - the factory will have full control
over the class loader
- better support and integration with existing java beans.

All that while maintaining full compatibility with ant1.4 - all
the code here is just an additional set of hooks, with the previous
behavior preserved and remaining as default.

Eventually this can be refactored into a standalone component
that will implement a task engine for ant1.x.


git-svn-id: http://svn.apache.org/repos/asf/ant@271572 13f79535-47bb-0310-9956-ffa450edef68