Since nobody objects, this is the 'project helper modularization' commit.
The default is ProjectHelperImpl, which have the original ( SAX1 ) implementation.

Most of the XML-related code has been removed, the only remaining problem
is the AttributeList that is passed to the introspection helper ( which is
XML and SAX1 specific ). It's easy to pass this by using a AttributeListImpl,
and hopefully the introspection helper will be made less SAX1-specific someday.


git-svn-id: http://svn.apache.org/repos/asf/ant@271780 13f79535-47bb-0310-9956-ffa450edef68