org.apache.tools.ant.taskdefs.cvslib.CvsTagDiff
The goal of this task is to give a list of differences between two tags or
dates. It produces an xml report like the changelog report of the CvsChangeLog
task. The output gives the list of modified files, removed files and new files.

The proposed task name is cvstagdiff.

PR: 8086
Submitted by: fred@castify.net (Frederic Lavigne)


git-svn-id: http://svn.apache.org/repos/asf/ant@272429 13f79535-47bb-0310-9956-ffa450edef68