PR 8793; Need to synchronize core methods in Project.
added sync attr to the various setters of Project state. We will take a fractional performance hit on this, even when we arent in parallel mode, which is why I use the synchronized attribute on the setProperty methods; that is ever so slightly faster than syncing on objects inside a method.


git-svn-id: http://svn.apache.org/repos/asf/ant@272849 13f79535-47bb-0310-9956-ffa450edef68