Allow to specify a given factory implementation for the transformer (xsltc, xalan, saxon...) and to specify processor specific settings.
TraxLiaison has been completely refactored because the factory
needed to be created once everything was set up.

I don't think it is final state since it might be better to specify factory attributes
and element in a <factory> element rather than how it is now.

XSLTProcess starts to bevery difficult to read with all the inner classes as well...


git-svn-id: http://svn.apache.org/repos/asf/ant@273089 13f79535-47bb-0310-9956-ffa450edef68