A small refactoring, getting to a more 'real' interface for the dynamic properties.
This is not final, of course.

PropertyHelper will implement all property manipulation. Eventually in
ant1.6 the static methods in various places should just wrap and
call this. In 1.6 the property storage can also be migrated to this
class.

It should be possible by a task or embedding application to replace
the whole mechanism - no need for discovery on this one, it can be
done by a task.

The PropertyInterceptor will be used to plug different sources
for the property. I'm thinking to just have tasks/types implementing
this interface automatically get registered - it's the simpler
solution for tasks developers ( I think ).


git-svn-id: http://svn.apache.org/repos/asf/ant@273107 13f79535-47bb-0310-9956-ffa450edef68