Finally, make all AntHandlers 'stateless' and make sure all the state is in AntContext.
Started the cleanup process.

The reason for 'stateless' model is not performance, but having a cleaner
model. It is very hard to track things on the stack or in different
handlers.

The next major step will be to try to reorganize the DataType/Task/UnknownTask
creation.

So far all features ( sax2, import, dynamic prop, classloader ) seem to work
fine with 1.5 - at least on all projects I'm working on.


git-svn-id: http://svn.apache.org/repos/asf/ant@273164 13f79535-47bb-0310-9956-ffa450edef68