Added a bit more info :-)
The next major step for this will be a set of hooks to allow ant to
be more JMX friendly - one very straightforward way to embed ant
is by using MBeans to represent the components.

In addition ant should be able to use MBeans as tasks and in
the build process.


git-svn-id: http://svn.apache.org/repos/asf/ant@273165 13f79535-47bb-0310-9956-ffa450edef68