Check in 11644. Thanks Nicola.
XMLDOC is a typedes that can be used in the embed proposal.
It loads an xml file as a DOM, thus making it available to jxpath.

Note:
This is different from the XML task that loads each element
as a property. Eventually we should merge the 2 - after 'embed'
is proposed for the main branch and if it is accepted.

Note2: It would be interesting to explore the other side of
jxpath, i.e. setting. And then saving the xml file.


git-svn-id: http://svn.apache.org/repos/asf/ant@273217 13f79535-47bb-0310-9956-ffa450edef68