Few small changes and comments.
I would like to extend it a bit more in the future - to support
properties like "${property}" without a string conversion.

In other workds, if a dynamic property is used ( which return an
object ), the object should be used as such in the introspection.
That would greatly simplify tasks and allow more flexibility.


git-svn-id: http://svn.apache.org/repos/asf/ant@273433 13f79535-47bb-0310-9956-ffa450edef68