Added a callback for setProperty. This would allow different property stores to implement their own policy.
This feature will not work in ant15 ( for now ).

Also changed Project to Object - and use it as a context. This makes
the callback more generic and the hooks less dependent of ant.

The only major issue on dynamic properties for me is the JNDI stuff.
I think the Context interface could be a better abstraction, but
on the other side it is much more complex. I also want to finalize
the JNDI property source and abstract the properties via JNDI
so tasks can be created without any dep on ant, but still accessing
properties and refs.


git-svn-id: http://svn.apache.org/repos/asf/ant@273434 13f79535-47bb-0310-9956-ffa450edef68