That's the big change.
I commented out for now the special case that does imediate
creation of task, and changed the behavior to all-lazy.

This seems to pass the gump test, and doesn't seem to
create any major problem. Given that lazy creation was allways
a possibility, I would think it is a task bug if a task that can't deal with
that.

After this settles down I'll eliminate the latest special case for
nested elements ( now it'll all be dealt only by UnknownElement ).

Also, added more code to deal with namespaces. The logic should
be in UnknownElement/RuntimeConfigrable - we should just pass
the info ( in the case that an XML source is used )


git-svn-id: http://svn.apache.org/repos/asf/ant@273438 13f79535-47bb-0310-9956-ffa450edef68