Modest contribution to help Stefan if he wants to test for Forrest.
I doubt this is complete and there should probably be some tweaks
to do as it will remove some br formatting (and I did not test yet under
Forrest, only from the top of my head). Not perfect, but this should
somewhat ease the conversion by doing most of the borring work.

cya


git-svn-id: http://svn.apache.org/repos/asf/ant@273608 13f79535-47bb-0310-9956-ffa450edef68