Add a method to support configuration using SAX2 attributes.
Changed ( at least for few now ) the default helper to the SAX2
processor.
The original processor can still be used using the system property,
and we can revert if we have see problems.

Description4 still fails - I'll add try to fix it later, I
don't think it's a big show-stopper.


git-svn-id: http://svn.apache.org/repos/asf/ant@273656 13f79535-47bb-0310-9956-ffa450edef68