Make it compile with HEAD.
After the change to ProjectHelper2, [embed] shouldn't be
used with ant1.6, only with ant1.5 as a mechanism to use the
new tasks and semantics with the stable branch.

The build file for Ant1.6 should be the same ( and work the same)
as the build file for ant1.5.1+embed.
This is not the case at the moment, because embed supports import
and few other features that are not yet in the main branch.

Import and the system loader can be supported now as regular tasks.
( preferably moved into HEAD, but they can work as an antlib )

Dynamic properties and ProjectComponentFactory still need to be
discussed and merged to HEAD ( or replaced with whatever gets
in the HEAD ).


git-svn-id: http://svn.apache.org/repos/asf/ant@273665 13f79535-47bb-0310-9956-ffa450edef68