Add the <classloader> task.
It'll create a new loader or add paths to an existing loader.

The main use is for the core loader - used to load ant regular
and optional tasks.

This is similar with embed, but 90% of the code is not needed for
ant1.6 ( due to the delayed task creation ).

For both import and classloader I changed the name of the class
so embed can still compile.


git-svn-id: http://svn.apache.org/repos/asf/ant@273688 13f79535-47bb-0310-9956-ffa450edef68