Few small changes. If the core loader is set and no explicit classpath is present, use it.
I think this can be further cleaned up to use the right parent loader
etc. But probably the best would be to just use the loaderRef.
Or just deprecate Definer completely - when we have a good antlib solution

Also added getters ( it's nice to have access to info - especially with
dynamic properties :-)


git-svn-id: http://svn.apache.org/repos/asf/ant@273720 13f79535-47bb-0310-9956-ffa450edef68