Finally fix the jspc test failure.
This is just the first round, it can be improved - not sure what's
the best way to deal with classpath specific issues in path.

The problem is that CLASSPATH ( as given in the system property,
and added in concatSystemPath ) may contain relative paths,
and they are relative to user.dir, not the project basedir.

Since gump is using relative paths, the launched java didn't find
the classes it needed, returning the strange -1 error.

Of course, the test suite could be more informative too :-)

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/ant@273727 13f79535-47bb-0310-9956-ffa450edef68