Cut&paste of all code related to task/types from Project.
This is _not_ an antlib, just a mechansim to make the task/type
management pluggable and simplify Project ( which will just delegate
for backward compat. ).

Please review - and -1 ( or just remove/change ) if you think this
is bad.

The actual details of the plugin mechanisms are not perfect -
we need to settle on them before 1.6 is released ( same for PropertyHelper).

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/ant@273745 13f79535-47bb-0310-9956-ffa450edef68