Few changes: - implement the BuildLogger - so it can be used with -logger
- few hacks/changes to support logging messages to stdout/stderr without
going to infinite loop. The trick is to create the logger with stdout
beeing the "real" one. Just to be sure it also logs with stdout set to
the real one. This could be done more cleanely probably - with some
changes in core.

- less verbosity ( task started/stoped doesn't need info, etc ).

- the log category is based on the real task name, combine with the
target. That means you can arbitrarily enable debug on tasks and
to tasks that are in particular target.



PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/ant@273772 13f79535-47bb-0310-9956-ffa450edef68