Allow <exec>'s error and output streams to be redirected independently
This is primarily to address bug 7330 but it includes a few other changes.

You can now use both output and outputproperty. While it was not documented,
previously, these used to be mutually exclusive. A simple TeeOutptuStream
is used to write the value to both a property and a file, if required.

The LogStreamHandler is no longer used by Exec. The main thing it did over
the PumpStreamHandler was to close the output streams when the stream
handler's stop method was called. This needed to be done in the
PumpStreamHandler in any case or the output file would be left open.
The PumpStreamHandler can now close its streams on stop() although the
default is still to leave them open

There are new attributes "error" and "errorProperty" which if present
separate output and error streams. If not present the current
behaviour is unchanged.

PR:	7330


git-svn-id: http://svn.apache.org/repos/asf/ant@274003 13f79535-47bb-0310-9956-ffa450edef68