Plug the namespace uri. One way or another - we'll need it.
Also fix the qname - we need to use the localname.

Originally used qname to avoid breaking build files that used : ( they still
needed to declare a namespace for that previx - otherwise SAX would complain ).
( not sure about the last part - I think some parsers, when validation is off, will
allow : to be used and pass whatever was used as qname ).


git-svn-id: http://svn.apache.org/repos/asf/ant@274547 13f79535-47bb-0310-9956-ffa450edef68