Revert part of the commit that made tasks use Execute#isFailure.
On OpenVMS the Java VM will exit with a return code of 0 on success,
even though this is going to signal a failure for the OS.  All tasks
that spawn new VMs have now been reverted to explicitly check for
retCode != 0 instead of Execite.isFailure(retCode).

We could as well introduce Execute#isJavaFailure or something like
this in case future VM's on OpenVMS will start to do the right thing.


git-svn-id: http://svn.apache.org/repos/asf/ant@275005 13f79535-47bb-0310-9956-ffa450edef68