1. new ildasm command to disassemble .il files. So that you can disassemble the output of tlbimport, patch it and recompile it. Yes, that is the official way to get C arrays marshalled into .net
2. patch to ilasm so that the default fileset is ignored when <src> elements are added and srcDir attr is undefined

3. tests!

ildasm has duff resource handling. It autogenerates the resources in the dir it runs, you cannot predict or control their names. So we CD to output dir, that is the dir of destFile unless an explicit res dir is provided. Ugly, ugly ugly.

And to think that they claim .net is better on account of the ease of development.


git-svn-id: http://svn.apache.org/repos/asf/ant@275190 13f79535-47bb-0310-9956-ffa450edef68