Make failure to display the splash screen non-fatal.
I've tested this with JDK 1.4 and the java.awt.headless system
property set to true as well as a Unix box running without an X
server.  While the former throws an Exception in the constructor of
SplashScreen, the later throws an Error in the constructor of
ImageIcon.

PR: 11482

Fix a copyright year mistake introduced in revision 1.9 at the same time.


git-svn-id: http://svn.apache.org/repos/asf/ant@276038 13f79535-47bb-0310-9956-ffa450edef68