Made web.xml non optional again, as it was hiding too many bugs (e.g war files for tomcat 5), but added an attribute to make it optional.
More tests, especially for updates. One test is disabled because it fails. The task fails if you are pulling in a web.xml on an existing web.xml file via a fileset if update=false, because the fileset version checking is hiding the inclusion of the web.xml file from the war task -its being dropped before we notice. This is not a BC problem. Maybe I should always make web.xml mandatory to stop this behavior arising.

git-svn-id: http://svn.apache.org/repos/asf/ant@468568 13f79535-47bb-0310-9956-ffa450edef68